# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "JAXRS RESTEasy Demo",
    "version" : "1.0.0",
    "title" : "Swagger Demo"
  },
  "host" : "localhost:8080",
  "basePath" : "/bookstore/rest",
  "tags" : [ {
    "name" : "genres"
  } ],
  "schemes" : [ "http" ],
  "paths" : {
    "/genres" : {
      "get" : {
        "tags" : [ "genres" ],
        "operationId" : "listAll",
        "parameters" : [ {
          "name" : "start",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "max",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Genre"
              }
            },
            "headers" : { }
          }
        }
      },
      "post" : {
        "tags" : [ "genres" ],
        "operationId" : "create",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Genre"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    },
    "/genres/{id}" : {
      "get" : {
        "tags" : [ "genres" ],
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "put" : {
        "tags" : [ "genres" ],
        "operationId" : "update",
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Genre"
          }
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      },
      "delete" : {
        "tags" : [ "genres" ],
        "operationId" : "deleteById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "pattern" : "[0-9][0-9]*",
          "format" : "int32"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        }
      }
    }
  },
  "definitions" : {
    "Genre" : {
      "type" : "object",
      "properties" : {
        "genre" : {
          "type" : "string"
        },
        "genreId" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    }
  }
}